fix: frontendの構成を見直し
